Add my first homework!
Hereby I have presented u with my first Python homework.
This is a small code to work with nucleotide RNA or DNA sequences. It can make transcription, reversion, complement, and complement with reversion. Pls do not try to mix RNA and DNA sequences as well as unusual nucleotides (not ATGC) — it will swear and spear you.