Added my second homework
The following program contains only definitions of functions to filter *.fastq output for GC content, length, and quality by the mean phred33 score. To use it, call the _main_ function with two paths and ur favourite boundaries in int, float, or tuple format. Be aware, there is no foolproof mechanism!